requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.20
langchain==0.3.0
langchain-openai==0.2.0
langchain-community==0.3.0
langchain-core==0.3.0
pypdf==6.1.0
faiss-cpu
pydantic==2.9.0
python-dotenv==1.0.1
pytest==8.3.0
pytest-asyncio==0.24.0
httpx==0.27.0
openai==1.54.0
requests==2.32.0

